function/requirements.txt
google-cloud-pubsub
google-cloud-storage
google-cloud-aiplatform
google-cloud-firestore
vertexai
